fix(swagger): Move nolint,deadcode instructions to own file
Closes https://github.com/ory/docs/pull/279
