--- identity/doc.go
@@ -1,9 +1,0 @@
-package identity
-
-// swagger:parameters createIdentity upsertIdentity
-// nolint:deadcode,unused
-type swaggerParametersCreateIdentity struct {
-	// in: body
-	// required: true
-	Body Identity
-}
